rubocop changes

+++ app/models/spree/bookkeeping_document.rb
@@ -1,11 +1,11 @@
 module Spree
   class BookkeepingDocument < ActiveRecord::Base
-    PERSISTED_ATTRS = [
-      :firstname,
-      :lastname,
-      :email,
-      :total,
-      :number
+    PERSISTED_ATTRS = %i[
+      firstname
+      lastname
+      email
+      total
+      number
     ]
 
     # Spree::BookkeepingDocument cares about creating PDFs. Whenever it needs to know
@@ -16,17 +16,17 @@ module Spree
     #
     belongs_to :printable, polymorphic: true
     belongs_to :store
-    belongs_to :print_invoice_setting, foreign_key: :setting_id, class_name: "Spree::PrintInvoiceSetting"
+    belongs_to :print_invoice_setting, foreign_key: :setting_id, class_name: 'Spree::PrintInvoiceSetting'
     validates :printable, :template, presence: true
-    validates *PERSISTED_ATTRS, presence: true, if: -> { self.persisted? }
+    validates(*PERSISTED_ATTRS, presence: true, if: -> { persisted? })
     scope :invoices, -> { where(template: 'invoice') }
 
     before_create :copy_view_attributes
     after_save :after_save_actions
 
-  def self.ransackable_attributes(auth_object = nil)
-    ["created_at", "email", "firstname", "lastname", "number", "printable_id", "printable_type", "template"]
-  end
+    def self.ransackable_attributes(_auth_object = nil)
+      %w[created_at email firstname lastname number printable_id printable_type template]
+    end
 
     # An instance of Spree::Printable::#{YourModel}::#{YourTemplate}Presenter
     #
@@ -145,9 +145,7 @@ module Spree
     # Renders and stores it if it's not yet present.
     #
     def send_or_create_pdf
-      unless File.exist?(file_path)
-        File.open(file_path, 'wb') { |f| f.puts render_pdf }
-      end
+      File.open(file_path, 'wb') { |f| f.puts render_pdf } unless File.exist?(file_path)
 
       IO.binread(file_path)
     end
